+++ src/Component/OpenApi3/Tests/fixtures/use-cacheable-supports-method/expected/Normalizer/SchemaNormalizer.php
@@ -39,7 +39,7 @@ class SchemaNormalizer implements DenormalizerInterface, NormalizerInterface, De
             return new Reference($data['$recursiveRef'], $context['document-origin']);
         }
         if (\array_key_exists('floatProperty', $data) && \is_int($data['floatProperty'])) {
-            $data['floatProperty'] = (double) $data['floatProperty'];
+            $data['floatProperty'] = (float) $data['floatProperty'];
         }
         if (\array_key_exists('stringProperty', $data)) {
             $object->setStringProperty($data['stringProperty']);
